Fixes for Renderer changes
Move router_to_mux to app folder as that kind of functionality belongs
there at the moment. Fix tests to deal with changes to Renderer
interface that happened when render package was fleshed out. Remove
Serves references and replace them with RespondsTo calls as we can use
the embedded Checker calls instead of the unnecessary Serves calls