Move Contextable, Requesish Renderable into thoreni
Thoreni Shared values are moving to the root namespace, so each package
can only grab what they need, without needing a types or shared
namespace in thoreni. Requestish now shares Contextable with the actual
Request at the moment, but I need to migrate to using helpers or using
the actual http.Request in the long and short run.