Fix backwards err check forced stdout logging
I had an error check backwards, that forced the use of stdout as the
logging method because I was checking for the working directory, but had
the error check backwards. As you can almost always get the working
directory and my first check was for the working directory, the code
wouldn't ever get into further checks. Because each error checking
message was fairly unique it wasn't that hard to find which check was
backwards.